淞沪会战推演：真实地形/兵力 ruleset + 多点反击夺占演示
- 修复 Cell 坐标 u8 溢出 bug（15x21=315 格>255，88师(12,17)=269 溢出成 13 被错渲染到长江上）。demo 棋盘小未触发，真实大战场暴露。
- songhu.toml：15x21 长江/虹口/罗店/大场/四行仓库真实战局
- main.rs 淞沪演示：88师北上夺渡口、87师夺大场、炮兵占闸北、地方军夺罗店 → 国军4点胜
- 数据驱动移动链（move_points=2 连续 step）+ 确定性回放全通过

+++ src/event.rs
@@ -11,7 +11,7 @@
 use serde::{Deserialize, Serialize};
 
 /// 棋盘坐标（扁平索引，row*cols+col）
-pub type Cell = u8;
+pub type Cell = u16;
 
 /// 单位 ID
 pub type UnitId = u8;
